creating the instrumentItem logic based sampler
is still not ready: i need yet to hack the tone sampler (or maybe make a
prototype.function to it, so it is available to others)

I need to make the actual instrumentItem instancing protocol, that would
replace the current channel-based instancing protocol... the instrument
item is aimed to create jquery elements and all. They should be flexible
enough to allow synths aswell. The idea is that an instrumentItem is
created once a samplerDatabaase item is selected and not to just create
them all. The idea is to have a really big instruments database that I
can create and delete in realtime, in a performance

Also I need to implement the great new sliders I made in the
mscompose895 project.

Also need to